requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
bs4==0.0.2
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cssselect2==0.9.0
defusedxml==0.7.1
Django==6.0.2
djangorestframework==3.16.1
idna==3.11
pillow==12.1.1
pycparser==3.0
requests==2.32.5
reportlab==4.4.3
soupsieve==2.8.3
sqlparse==0.5.5
tinycss2==1.5.1
typing_extensions==4.15.0
urllib3==2.6.3
webencodings==0.5.1
